+++ crates/apps/rugix-ctrl/generate-json-schema.sh
@@ -10,6 +10,7 @@ cp generated/rugix_ctrl.state.StateConfig.schema.json ../../../schemas/rugix-ctr
 cp generated/rugix_ctrl.system.SystemConfig.schema.json ../../../schemas/rugix-ctrl-system.schema.json
 cp generated/rugix_ctrl.config.Config.schema.json ../../../schemas/rugix-ctrl-config.schema.json
 cp generated/rugix_ctrl.daemon.DaemonConfig.schema.json ../../../schemas/rugix-ctrl-daemon.schema.json
+cp generated/rugix_ctrl.component.ComponentDeclaration.schema.json ../../../schemas/rugix-ctrl-component.schema.json
 cp generated/rugix_ctrl.output.SystemInfoOutput.schema.json ../../../schemas/rugix-ctrl-output-info.schema.json
 cp generated/rugix_ctrl.output.ComponentsOutput.schema.json ../../../schemas/rugix-ctrl-output-components.schema.json
 cp generated/rugix_ctrl.output.ComponentsCheckOutput.schema.json ../../../schemas/rugix-ctrl-output-components-check.schema.json